Crate doc_index

This is the index page for the documentation to the package rly. The package consists of three crates:

lex for creating Lexers capable of parsing an input into Token’s.

parse for creating Parsers that can parse a stream of Tokens into an Ast.

rly_common containing code common to both lex and parse. Most significant for the end-user are the types ErrorData and LineOffset.